+++ app/src/main/java/pl/pjatk/squashme/model/custom/MatchHistory.java
@@ -1,4 +1,4 @@
-package pl.pjatk.squashme.model.complex;
+package pl.pjatk.squashme.model.custom;
 
 public class MatchHistory {
