+++ app/src/main/java/pl/pjatk/squashme/service/MatchService.java
@@ -6,9 +6,8 @@ import java.util.Optional;
 import io.reactivex.rxjava3.core.Single;
 import pl.pjatk.squashme.model.Match;
 import pl.pjatk.squashme.model.MatchWithPlayers;
-import pl.pjatk.squashme.model.complex.MatchHistory;
-import pl.pjatk.squashme.model.complex.TournamentMatchSimple;
-import pl.pjatk.squashme.model.complex.TournamentResults;
+import pl.pjatk.squashme.model.custom.MatchHistory;
+import pl.pjatk.squashme.model.custom.TournamentMatchSimple;
 
 public interface MatchService {
