+++ app/src/main/java/pl/pjatk/squashme/service/MatchServiceImpl.java
@@ -9,8 +9,8 @@ import io.reactivex.rxjava3.core.Single;
 import pl.pjatk.squashme.dao.MatchDao;
 import pl.pjatk.squashme.model.Match;
 import pl.pjatk.squashme.model.MatchWithPlayers;
-import pl.pjatk.squashme.model.complex.MatchHistory;
-import pl.pjatk.squashme.model.complex.TournamentMatchSimple;
+import pl.pjatk.squashme.model.custom.MatchHistory;
+import pl.pjatk.squashme.model.custom.TournamentMatchSimple;
 
 public class MatchServiceImpl implements MatchService {
