+++ app/src/main/java/pl/pjatk/squashme/service/TournamentService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import io.reactivex.rxjava3.core.Single;
 import pl.pjatk.squashme.model.Tournament;
-import pl.pjatk.squashme.model.complex.TournamentResults;
+import pl.pjatk.squashme.model.custom.TournamentResults;
 
 public interface TournamentService {
